Check if key images already in mempool
Partially solves issue:
https://github.com/moneroexamples/openmonero/issues/48

It checks if a tx that you are submitting contains key images
that are present in mempool, and print outs an appropiate
error message.

Ideally, differnt outputs should be used by the frontend in this case.
But for now, we have to just wait till our past txs get confirmed and disappear from the mempool.